+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [0.9.0] / 2026-04-06
+## [0.9.0] / 2026-04-06 - 2026-04-10
 ### Features
 - Support `Revit 2027` and `Revit Preview`.
 - Support `net10.0-windows` framework.